+++ src/execution/FetchStrategy.js
@@ -13,8 +13,11 @@ export default {
       case 'FETCH_FROM_NETWORK':
         return 'network only';
 
-      case 'FETCH_FROM_NETWORK_AND_NO_CACHE':
-        return 'network only and no caching';
+      case 'FETCH_FROM_NETWORK_AND_SKIP_CACHING':
+        return 'network only and skip caching';
+
+      case 'FETCH_FROM_NETWORK_AND_SKIP_CACHING_AND_CACHE_UPDATE':
+        return 'network only and skip caching and cache update';
     }
   },
   ...Object.freeze({
@@ -23,7 +26,7 @@ export default {
     FetchFromCacheAndNetwork: 'FETCH_FROM_CACHE_AND_NETWORK',
     FetchFromNetwork: 'FETCH_FROM_NETWORK',
     FetchFromNetworkAndRecreateCache: 'FETCH_FROM_NETWORK_AND_RECREATE_CACHE',
-    FetchFromNetworkAndNoCache: 'FETCH_FROM_NETWORK_AND_NO_CACHE',
-    FetchFromNetworkAndNoCacheNoCacheUpdates: 'FETCH_FROM_NETWORK_AND_NO_CACHE_NO_CACHE_UPDATES'
+    FetchFromNetworkAndSkipCaching: 'FETCH_FROM_NETWORK_AND_SKIP_CACHING',
+    FetchFromNetworkAndSkipCachingAndCacheUpdate: 'FETCH_FROM_NETWORK_AND_SKIP_CACHING_AND_CACHE_UPDATE'
   })
 };